Adding Riposte and fixing bug (jpg frames being created)
