--- Info to run.txt
@@ -1,12 +1,0 @@
-You need to put a token for a bot in to bad apple discord player program so that it can connect
-
-You should put the video to frames program in a folder as it puts every frame in its own directory
-
-You might have to change some paths in the files
-
-Some modules are needed to run the programs, they are:
-	- Discord
-	- Time
-	- Pillow
-	- OpenCV
-
