+++ requirements.txt
@@ -1,0 +1,3 @@
+riposte==0.4.1
+discord==1.0.1
+Pillow==8.1.0
